Removing UtilityService class.
