+++ FoodOrderingApp-api/src/main/java/com/upgrad/FoodOrderingApp/api/controller/CustomerController.java
@@ -2,7 +2,6 @@ package com.upgrad.FoodOrderingApp.api.controller;
 
 import com.upgrad.FoodOrderingApp.api.model.*;
 import com.upgrad.FoodOrderingApp.service.businness.CustomerService;
-import com.upgrad.FoodOrderingApp.service.businness.UtilityService;
 import com.upgrad.FoodOrderingApp.service.entity.CustomerAuthEntity;
 import com.upgrad.FoodOrderingApp.service.entity.CustomerEntity;
 import com.upgrad.FoodOrderingApp.service.exception.AuthenticationFailedException;
@@ -24,11 +23,7 @@ import java.util.UUID;
 public class CustomerController {
 
     @Autowired
-    private UtilityService utilityService;
-
-    @Autowired
     private CustomerService customerService;
-
 
     /**
      * @param signupCustomerRequest
@@ -52,7 +47,7 @@ public class CustomerController {
         customerEntity.setPassword(signupCustomerRequest.getPassword());
         customerEntity.setContactnumber(signupCustomerRequest.getContactNumber());
 
-        utilityService.isValidSignupRequest(customerEntity);
+        customerService.isValidSignupRequest(customerEntity);
 
         final CustomerEntity createdCustomerEntity = customerService.saveCustomer(customerEntity);
 
@@ -70,7 +65,7 @@ public class CustomerController {
     public ResponseEntity<LoginResponse> login(@RequestHeader("authorization") final String authorization)
             throws AuthenticationFailedException {
         // Basic authentication format validation
-        utilityService.isValidAuthorizationFormat(authorization);
+        customerService.isValidAuthorizationFormat(authorization);
 
         //Separating the username and password after decoding using Base64 decoder
         byte[] decoded = Base64.getDecoder().decode(authorization.split("Basic ")[1]);
@@ -119,11 +114,11 @@ public class CustomerController {
             @RequestHeader("authorization") final String authorization,
             @RequestBody(required = false) UpdateCustomerRequest updateCustomerRequest)
             throws UpdateCustomerException, AuthorizationFailedException {
-        utilityService.isValidUpdateCustomerRequest(updateCustomerRequest.getFirstName());
+        customerService.isValidUpdateCustomerRequest(updateCustomerRequest.getFirstName());
 
         //Access the accessToken from the request Header
         String accessToken = authorization.split("Bearer ")[1];
-        utilityService.validateAccessToken(accessToken);
+        customerService.validateAccessToken(accessToken);
 
         CustomerEntity customerEntityToBeUpdated = customerService.getCustomer(accessToken);
 
@@ -154,11 +149,11 @@ public class CustomerController {
         String oldPassword = updatePasswordRequest.getOldPassword();
         String newPassword = updatePasswordRequest.getNewPassword();
 
-        utilityService.isValidUpdatePasswordRequest(oldPassword, newPassword);
+        customerService.isValidUpdatePasswordRequest(oldPassword, newPassword);
 
         //Access the accessToken from the request Header
         String accessToken = authorization.split("Bearer ")[1];
-        utilityService.validateAccessToken(accessToken);
+        customerService.validateAccessToken(accessToken);
 
         CustomerEntity customerEntityToBeUpdated = customerService.getCustomer(accessToken);
