+++ FoodOrderingApp-service/src/main/java/com/upgrad/FoodOrderingApp/service/businness/CustomerService.java
@@ -14,13 +14,13 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.ZonedDateTime;
+import java.util.Base64;
 import java.util.UUID;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @Service
 public class CustomerService {
-
-    @Autowired
-    private UtilityService utilityService;
 
     @Autowired
     private CustomerAuthDao customerAuthDao;
@@ -30,6 +30,111 @@ public class CustomerService {
 
     @Autowired
     PasswordCryptographyProvider passwordCryptographyProvider; //Provides coding and decoding for the password
+
+    //To validate the Authorization format
+    public boolean isValidAuthorizationFormat(String authorization) throws AuthenticationFailedException {
+        try {
+            byte[] decoded = Base64.getDecoder().decode(authorization.split("Basic ")[1]);
+            String decodedAuth = new String(decoded);
+            String[] decodedArray = decodedAuth.split(":");
+            String username = decodedArray[0];
+            String password = decodedArray[1];
+            return true;
+        } catch (ArrayIndexOutOfBoundsException exc) {
+            throw new AuthenticationFailedException("ATH-003", "Incorrect format of decoded customer name and password");
+        }
+    }
+
+    //To validate Customer update request
+    public boolean isValidUpdateCustomerRequest(String firstName) throws UpdateCustomerException {
+        if (firstName == null || firstName == "") {
+            throw new UpdateCustomerException("UCR-002", "First name field should not be empty");
+        }
+        return true;
+    }
+
+    //To validate the password Update Request.
+    public void isValidUpdatePasswordRequest(String oldPassword, String newPassword) throws UpdateCustomerException {
+        if (oldPassword == null || oldPassword == "") {
+            throw new UpdateCustomerException("UCR-003", "No field should be empty");
+        }
+        if (newPassword == null || newPassword == "") {
+            throw new UpdateCustomerException("UCR-003", "No field should be empty");
+        }
+        return;
+    }
+
+    public boolean isValidSignupRequest(CustomerEntity customerEntity) throws SignUpRestrictedException {
+        if (customerEntity.getFirstName() == null || customerEntity.getFirstName() == "") {
+            throw new SignUpRestrictedException("SGR-005", "Except last name all fields should be filled");
+        }
+        if (customerEntity.getPassword() == null || customerEntity.getPassword() == "") {
+            throw new SignUpRestrictedException("SGR-005", "Except last name all fields should be filled");
+        }
+        if (customerEntity.getEmail() == null || customerEntity.getEmail() == "") {
+            throw new SignUpRestrictedException("SGR-005", "Except last name all fields should be filled");
+        }
+        if (customerEntity.getContactnumber() == null || customerEntity.getContactnumber() == "") {
+            throw new SignUpRestrictedException("SGR-005", "Except last name all fields should be filled");
+        }
+        return true;
+    }
+
+    private boolean isPasswordWeak(String password) {
+        // Regex to check valid password.
+        String regex = "^(?=.*[0-9])"
+                + "(?=.*[A-Z])"
+                + "(?=.*[#@$%&*!^])"
+                + "(?=\\S+$).{8,20}$";
+
+        // Compile the ReGex
+        Pattern p = Pattern.compile(regex);
+
+        // Pattern class contains matcher() method
+        // to find matching between given password
+        // and regular expression.
+        Matcher m = p.matcher(password);
+
+        // Return if the password
+        // matched the ReGex
+        return m.matches();
+    }
+
+    private boolean isEmailValid(String email) {
+        String regex = "^[\\w-_\\.+]*[\\w-_\\.]\\@([\\w]+\\.)+[\\w]+[\\w]$";
+        return email.matches(regex);
+    }
+
+    private boolean isContactNumberValid(String contactNumber) {
+        Pattern p = Pattern.compile("\\A[0-9]{10}\\z");
+        Matcher m = p.matcher(contactNumber);
+        return (m.find() && m.group().equals(contactNumber));
+    }
+
+    @javax.transaction.Transactional
+    public CustomerAuthEntity validateAccessToken(final String accessToken) throws AuthorizationFailedException {
+        CustomerAuthEntity customerAuthEntity = customerAuthDao.getCustomerAuthTokenByAccessToken(accessToken);
+
+        //Checking if Customer not logged In
+        if (customerAuthEntity == null) {
+            throw new AuthorizationFailedException("ATHR-001", "Customer is not Logged in.");
+        }
+
+        //Checking if customer is logged Out
+        if (customerAuthEntity.getLogoutAt() != null) {
+            throw new AuthorizationFailedException("ATHR-002",
+                    "Customer is logged out. Log in again to access this endpoint.");
+        }
+
+        final ZonedDateTime now = ZonedDateTime.now();
+
+        //Checking accessToken is Expired.
+        if (customerAuthEntity.getExpiresAt().compareTo(now) <= 0) {
+            throw new AuthorizationFailedException("ATHR-003",
+                    "Your session is expired. Log in again to access this endpoint.");
+        }
+        return customerAuthEntity;
+    }
 
     public CustomerEntity getCustomer(final String accessToken) throws AuthorizationFailedException {
         CustomerAuthEntity customerAuthEntity = customerAuthDao.getCustomerAuthTokenByAccessToken(accessToken);
@@ -46,19 +151,19 @@ public class CustomerService {
             throw new SignUpRestrictedException("SGR-001", "This contact number is already registered! Try other contact number");
         }
 
-        if (!utilityService.isValidSignupRequest(customerEntity)) {//Checking if is Valid Signup Request.
+        if (!isValidSignupRequest(customerEntity)) {//Checking if is Valid Signup Request.
             throw new SignUpRestrictedException("SGR-005", "Except last name all fields should be filled");
         }
 
-        if (!utilityService.isEmailValid(customerEntity.getEmail())) {//Checking if email is valid
+        if (!isEmailValid(customerEntity.getEmail())) {//Checking if email is valid
             throw new SignUpRestrictedException("SGR-002", "Invalid email-id format!");
         }
 
-        if (!utilityService.isContactNumberValid(customerEntity.getContactnumber())) {//Checking if Contact is valid
+        if (!isContactNumberValid(customerEntity.getContactnumber())) {//Checking if Contact is valid
             throw new SignUpRestrictedException("SGR-003", "Invalid contact number!");
         }
 
-        if (!utilityService.isPasswordWeak(customerEntity.getPassword())) {//Checking if Password is valid.
+        if (!isPasswordWeak(customerEntity.getPassword())) {//Checking if Password is valid.
             throw new SignUpRestrictedException("SGR-004", "Weak password!");
         }
 
@@ -145,7 +250,7 @@ public class CustomerService {
             final String newPassword,
             CustomerEntity customerEntity) throws UpdateCustomerException {
 
-        if (!utilityService.isPasswordWeak(newPassword)) {//Checking if the Password is Weak.
+        if (!isPasswordWeak(newPassword)) {//Checking if the Password is Weak.
             throw new UpdateCustomerException("UCR-001", "Weak password!");
         }
